+++ include/gelex/estimator/gibbs/bayes_c.h
@@ -31,7 +31,7 @@ void BayesCKernel(
     double var_a{};
     const dvec inv_scaler = 1.0 / (cols_norm + sigma_e / sigma_a);
 
-    for (uint64_t i = 0; i < a.n_elem; ++i)
+    for (size_t i = 0; i < a.n_elem; ++i)
     {
         if (cols_var.at(i) == 0.0)
         {
